requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
coverage==6.4.1
drf-yasg==1.20.0
django-filter==22.1
whitenoise==6.2.0
python-dotenv==0.21.0